refactor: AdminSchoolManageView -> AdminSchoolManageListView 이름 변경
