+++ src/router/modules/admin/adminRoutes.ts
@@ -1,6 +1,6 @@
 import RootAdminView from '@/views/root/RootAdminView.vue';
 import AdminMyPageView from '@/views/admin/AdminMyPageView.vue';
-import AdminSchoolManageView from '@/views/admin/AdminSchoolManageView.vue';
+import AdminSchoolManageListView from '@/views/admin/school/AdminSchoolManageListView.vue';
 import AdminFestivalManageView from '@/views/admin/AdminFestivalManageView.vue';
 
 const adminRoutes = [
@@ -16,8 +16,8 @@ const adminRoutes = [
   },
   {
     path: '/admin/school',
-    name: 'adminSchoolManage',
-    component: AdminSchoolManageView,
+    name: 'adminSchoolManageListView',
+    component: AdminSchoolManageListView,
   },
   {
     path: '/admin/festival',
